filestore: Change the key padding strategy.
Previously keys were unconditionally padded with two NUL bytes at the end prior
to hex encoding. This ensures no key is ever too short to have a directory and
a non-empty file path, but makes the store incompatible with Git object layout
when used as a CAS.

Now instead, pad with "-" (Unicode 45) after hex encoding. Not being a hex
digit, it can be stripped before decoding, and it sorts prior to any hex digit
so that keys can be sorted correctly without decoding.