Avert a spurious protobuf registration conflict warning.
The protobuf runtime check for registration conflicts contains what I consider
to be a bug: If two protobuf packages, even if their names are globally
distinct, generate output relative to source and use the same base filename,
they will trigger the warning. The warning will eventually become an error.

I tried fixing this by adding disjoint package declarations, but Go ignores
them. I also tried fixing this by adding proper go_package options, but that
does not fix the issue for filenames.  Nevertheless, both of those changes are
included here since they can't hurt.

The actual "fix" is to just rename the files. These two packages are internal
to this module anyway, and are fully disjoint. I should probably file a bug
about this, but the likelihood that it would get any response at all is very
low, and if it did that response would just be "WAI".