Deduplicate block size counters.
For small files it is common for the index to have only one extent, and for
that extent to have only a single block. Save a few bytes by not storing the
length multiple times:

- If an extent has only one block, inherit its size from the extent.
- If an index has only one extent, inherit its size from the total.

This change is implemented on encoding/decoding, so it is transparent to the
rest of the package.