Revert virtual ("view") file support.
I added this feature to support attaching snapshots in FUSE, but the
complications of hooking it up to the API were greater than I expected.
Moreover, the FUSE library dropped support for macOS, so while I'm able to keep
using it for now at an old commit, I expect it will eventually bitrot to the
point of unusability. I may fork it or do something simpler, but in either case
the API interactions will change. For now, I'm dropping the dead code.