Re-define the wire format in terms of protocol buffers.
I previously reverted this change in favour of using my own binpack package.
At the time, I did this because protobuf has a noisy dependency footprint, the
quality of the generated code was mediocre, and the encoding into wire format
is not deterministic over time.

Since then, however, the Go protobuf support has improved, it no longer pulls
in half the Google package ecosystem, and while the encoding order is still not
guaranteed, the Go implementation handles it sensibly for simple cases like the
schema used here (with no maps or extensions).