file: Fix handling of splitter configs.
The documentation claimed that split settings are persisted in storage, which
it used to be but no longer is. Fix that comment, and document that descendant
files created via the New method inherit the split settings of the parent, if
they do not have their own. Also, make that true.