wiretype: Make Stat depend less on Go API details.
The Go standard library packs permissions and mode bits into a vector in the
same way Unix does, but with a custom organization of the bits. Permissions are
still in the bottom 9 bits, while types and other bits are consistent, but
idiosyncratic to Go.

For storage, we now pack the standard 9 owner/group/other permission bits,
along with setuid/setgid/sticky bits, into a uint32 vector, but type
information is decoded into a separate enumeration. Since most files are
expected to be "regular" files, use 0 for that type so that we can omit it from
the wire in the common case.

Update the generated code.