file: Move exact data construction from package wiretype.
Remove the wiretype.NewIndex function, and move it to an unexported function of
the file package. This is mostly a renaming, but returns the internal fileData
type rather than the protobuf.

This is preparatory to giving a File a method to explicitly set its data
without an iterative copy. The problem with a standard copy is that it writes
the data in chunks, which cause block truncations at unexpected locations.
This is functionally harmless, but results in "dead" (unreferenced) blocks
being committed to the store.

This approach writes only and exactly those blocks to the store that are needed
to store the file data.