file: Set initial stat by pointer rather than value.
Rather than checking for an empty Stat value as the signal for persistence,
check for a nil pointer. This way, the caller can request persistence even if
they have no initial stat to populate. We still store the data by value, so
that we don't have to worry about mutation later.