Add a Bloom filter indexer for string keys (#4)
Add package index, implementing a simple Bloom filter for strings.

In order to prune data from the content-addressable store, we need a way to keep
track of which blobs belong to each root. Since hash keys are fairly long, it would
be expensive to keep them explicitly. For pruning, however, it's not essential that
we have a perfect record, as long as we can check whether a blob is used without
any risk of a false negative. This is a perfect use case for a Bloom filter.

Bloom filters do allow false positives, but we can tune the parameters to keep the rate
as low as necessary. A false positive will prevent us from pruning a blob that is not
actually in use, but as long as we don't do so too often, that won't matter.