wiretype: Add a special case for single-block data.
For the common case in small files, where file data consists of only one block,
store the key for that block directly rather than encoding it into an extent
and a block inside that extent. Update the encoding and decoding code to handle
this case specially.

Existing data with one single-block extent will continue to work as before, but
when writing such a file out, the encoding will be updated.