Add support for splitting extents.
When an extent contains zero-valued blocks, it can be split at those blocks
into multiple extents. Add a function that can split a single extent by
checking its blocks for these partition points.

To take advantage of this, splitBlobs has been updated to not store the
contents of zero-valued blobs.