Replace bytes.Reader with a custom reader.
The reblocking splitter was constructing a bytes.Reader for each blob to be
rescanned, and merging them via io.MultiWriter. This is a lot of allocation for
such a simple use. Replace that with a reader that just tracks where it is
along a slice of block slices and copies data to the reader.