Be forgiving of temporary errors.
The writeback thread would previously fail for any error in writing a blob.
This includes what I think are transient DNS errors in resolving GCS addresses,
when GCS is the target store.

Once the thread exits, it will not restart, and all subsequent requests through
the wrapper will fail. As an attempt to mitigate this, check for temporary DNS
errors and continue past them when they occur.