cachestore: cache keyspace for List and Len (#13)
Replace the non-existence cache with a complete keymap of the base store.  This
allows List and Len operations to be much faster for large remote storage like
GCS and S3 which do not have an efficient Len, and for which list operations
are expensive in both time and query operations.

Instead of a map, use a balanced binary tree and keep the non-existence cache
as the value for each key.