cachestore: ensure key map is always initialized
Since all the operations now require the keymap to be initialized, it is not
sufficient to only do this for List and Len. If the first operation on a new
cachestore is a Get for an existing key, an uninitialized keymap will report it
(incorrectly) as absent.

Fix that, and add a regression test.