Add package suffixed implementing a suffixed key store.
The prefixed key store has a deficiency for storage implementations that use
key prefixes for organization, because it groups the keys into a small number
of potentially large batches and breaks opportunities for parallelism.

Using a suffix instead means that scanning operations (Len, List) can't take
advantage of the prefix to terminate early; however in return they can make
better use of concurrency.