blob: require a Close method on the Store interface (#19)
Without having this method required, it is too easy for a wrapper
implementation to forget to expose the optional interface needed by the
delegate. At least if it is required, the implementation will have a place to
add the forwarding when it turns out to be missing. That is better than the
helper approach, which cannot see through the wrapping without help.