blob: redefine the CAS interface
Allow the caller to specify prefix and suffix fragments that are to be appended
to the content address for storage. Without this, an implementation that wants
to decorate content keys in an underlying store has to make two passes: One to
generate the content key, another to store the blob.

- Update the various storage implementations in this package and their tests.
- Update usage in the file and wiretype packages.