storage/wbstore: make Len and List respect the buffer store
Previously, the Len and List methods were delegated directly to the base store.
This is correct insofar as it reflects the "true" contents of the store, but if
a writeback store is terminated and restarted, the keys in the buffer that have
not yet been written back to the store will not be counted or listed.

Such keys are not lost; the writeback thread will still push them up to the
backing store eventually. However, this mixes poorly with a cachestore, since
the cache will only see the keys from the base store at the time of startup.