storage/hexkey: do not trim the shard prefix from its key
Previously, a key like 012345 with a shard length of 2 would be encoded as

    01/2345

that is, the shard prefix would be "cut" from its underlying key.  Now, the
original key is left alone, and the shard prefix is duplicated:

    01/012345

An empty key is still encoded as "-", when it occurs in a sharding:
The key "" with shard length 2 becomes "--/-".
The key "01" with shard length 1 becomes "0/01".