blob: rename Store to KV (#24)
We would like to use the name "Store" for a higher-level collection of
key-value buckets. As a first step, rename the "Store" interface, which itself
expresses a single key-value bucket, to "KV".

Future changes will introduce a new Store type that grants access to buckets.
In addition, I would like to break the nesting of CAS around KV, but for now I
have left the existing structure in place.