file: fix the Scan method to allow modification (#25)
Previously, the Scan was implicitly read-only, not holding locks except to
capture a view of the children of a file. This was safe, because the views of a
file already use locking where necessary, but makes a Scan unable to mutate the
state of a file.

Children of a file not already loaded into memory are loaded, but the Scan was
doing this unconditionally. For reading this does not matter, but when writing
it means we lose the connection between what the view provides and what the
parent believes the state of that child is.

So here there are three fixes:

1. Run the whole scan using the correct lock chaining, as Flush.
2. Only Open files not already present.
3. Update the parent for opened files mutated by the visitor.

Also update the tests.