Release v0.8.0
This release includes a substantial change to the blob package API.  The old
"Store" type was renamed to "KV" and a new top-level store type is introduced
allowing management of separate key spaces within the same storage artifact.

All the existing storage implementations in this module have been updated, and
although some smaller changes are yet to come, this is a good checkpoint.

Highlights:

- Update blob.Store and add blob.StoreCloser.
- Add common plumbing for flat-keyspace storage implementations.
- Extend the storetest package to exercise key spaces and substores.