Release v0.9.0
Another round of substantial API changes.

The main change is to update the Store interface to rename Keyspace to KV, and
to add a CAS method. In addition, the KV and CAS types are no longer nested.
Instead, both implement blob.KVCore, and then KV adds Put while CAS adds the
CASPut and CASKey methods. There is a one-way conversion provided to wrap a KV
into a CAS, but you have to express intent about which kind you want.

All the base storage implementations in this module are herewith updated.

Other changes:

- blob: add Closer as an explicit definition
- storage: remove the zipstore package
- storate/cachestore: implement the blob.Store interface
- storage/monitor: rework how config is plumbed

Maintenance:

- blob/memstore: update the package docs
- blob/memstore: update Opener to return a blob.StoreCloser
- storage/filestore: update Opener to return a blob.StoreCloser