storage/wbstore: make the buffer survive shutdown
When I converted this for the new Store API, I used prefix keys that were not
stable across runs of the program. This meant any keys in the buffer when the
system comes up were discarded. This is unnecessary, however: we can use the
same hash-key construction as for the flat key-value storage engines, based on
the naming path of substores and key spaces.

This also turns out to be a little less code, thanks to the monitor.