blob: add a Stat method to the KVCore interface
Now that CAS computation is all done at the leaves, it is once again useful to
have a way for the implementation to check presence before sending a full blob.

But instead of simply restoring the original Size method, which took one key at
a time, we can address both this and the "sync" case by having a variadic Stat
that also reports presence for multiple keys.

Update the default store implementations and tests.