blob: rework Stat into a plain existence check, rename to Has
As noted in 5f64dc6f, having to report the size had the side effect of making
this API inefficient for key space synchronization. Since I don't have a real
need for size right now anyway, rework the method as a plain existence check.
Update the test harness.

If size is needed later, that can be a separate method.