blob: update signature of List to return an iterator
Instead of a callback for keys, return an iterator the caller can pull.
To deal with errors, borrow a trick from @danderson, to report two values per
yield: In case of error, the List implementation yields a pair with a non-nil
error before terminating iteration. Otherwise, it yields a (key, nil) pair for
each valid key in the requested range.