storage/cachestore: allow read operations while listing
Previously, the need to update the key map during a Get prevented the body of a
List operation to access the contents of the store via Get, Has, etc.  This is
annoying, and unnecessary. Fix the state management so that you can do all
these things (as well as List itself) concurrently.