storage/wbstore: rework writeback storage
Instead of sharing a single flat buffer KV across all the stores, use a full
store and mirror the structure between base and buffer.

This gets rid of a bunch of fiddly prefix handling.  Instead of a single
write-behind worker, we now create one per KV.

However, now that each base KV has its own buffer KV, we need a different way
to monitor the size of the buffer. Add a hook to the Store that reports the
aggregate key count from all the buffer KVs.