Release v0.16.0
Breaking change: New interface for the codecs/encrypted package.

Instead of fixing a single cipher key, plumb in an AEAD constructor and a
keyring. The interface used here is meant to be satisfied eventually by the
github.com/creachadair/keyring package, but the codec supports both the
previous storage format and a new one with space for a key ID.

In codecs/encrypted:

  - extend the block format to allow a key ID
  - remove the random generator override
  - remove the Options type
  - change the constructor interface